Week 11-8-2023

+++ ResumableFunctions.Handler/DataAccess/ScanStateRepo.cs
@@ -1,24 +1,31 @@
 ﻿using Medallion.Threading;
 using Microsoft.EntityFrameworkCore;
+using ResumableFunctions.Handler.Core.Abstraction;
 using ResumableFunctions.Handler.DataAccess.Abstraction;
-using ResumableFunctions.Handler.InOuts;
+using ResumableFunctions.Handler.InOuts.Entities;
 
 namespace ResumableFunctions.Handler.DataAccess;
 
-//todo:it should be transient
 internal class ScanStateRepo : IScanStateRepo
 {
+    private readonly string _scanStateLockName;
     private readonly WaitsDataContext _context;
     private readonly IDistributedLockProvider _lockProvider;
+    private readonly IResumableFunctionsSettings _settings;
 
-    public ScanStateRepo(WaitsDataContext context, IDistributedLockProvider lockProvider)
+    public ScanStateRepo(
+        WaitsDataContext context,
+        IDistributedLockProvider lockProvider,
+        IResumableFunctionsSettings settings)
     {
         _context = context;
         _lockProvider = lockProvider;
+        _settings = settings;
+        _scanStateLockName = $"{_settings.CurrentServiceName}_ScanStateLock";
     }
     public async Task<bool> IsScanFinished()
     {
-        await using var lockScanStat = await _lockProvider.AcquireLockAsync("ScanStateLock");
+        await using var lockScanStat = await _lockProvider.AcquireLockAsync(_scanStateLockName);
         return await _context.ScanStates.AnyAsync() is false;
     }
 
@@ -32,8 +39,15 @@ internal class ScanStateRepo : IScanStateRepo
 
     public async Task<bool> RemoveScanState(int id)
     {
-        await using var lockScanStat = await _lockProvider.AcquireLockAsync("ScanStateLock");
+        if (id == -1) return true;
+        await using var lockScanStat = await _lockProvider.AcquireLockAsync(_scanStateLockName);
         await _context.ScanStates.Where(x => x.Id == id).ExecuteDeleteAsync();
         return true;
+    }
+
+    public async Task ResetServiceScanState()
+    {
+        await using var lockScanStat = await _lockProvider.AcquireLockAsync(_scanStateLockName);
+        await _context.ScanStates.Where(x => x.ServiceId == _settings.CurrentServiceId).ExecuteDeleteAsync();
     }
 }
